requirements.txt
kafka-python
prefect
pandas
clickhouse-driver